refactor(donation): cleanly separate donation logic for Razorpay integration
