--- .env
@@ -1,18 +1,0 @@
-# MONGO_URI=mongodb+srv://kaushiki123:ODtz4kAocaMQqcYn@cluster.mongodb.net/sitaare
-# PORT=5000
-# RAZORPAY_KEY_ID=rzp_test_LeKrj9wP3yNPoh
-# RAZORPAY_KEY_SECRET=urTw3WbAqBeA009UtdlTGGAY
-# REACT_APP_GOOGLE_MAPS_KEY=AIzaSyD0-UgDrMHZA4HRP_hOgFb0zz1JgZ8ZBpk
-
-
-# Razorpay Configuration
-# Get these from your Razorpay Dashboard → Settings → API Keys
-RAZORPAY_KEY_ID=rzp_test_o9wCW8xFj6ACL2
-RAZORPAY_KEY_SECRET=D4n6j9fO2R6FiMBVtKpSTY36
-
-# Database Configuration
-MONGO_URI=mongodb://localhost:27017/sitaare
-
-# Server Configuration
-PORT=5000
-JWT_SECRET=your_super_secret_jwt_key_here
