+++ server/controllers/donationController.js
@@ -1,0 +1,212 @@
+const Razorpay = require('razorpay');
+const crypto = require('crypto');
+const Donation = require('../models/Donation');
+const Beneficiary = require('../models/Beneficiary');
+const nodemailer = require('nodemailer');
+require('dotenv').config();
+
+const razorpay = new Razorpay({
+  key_id: process.env.RAZORPAY_KEY_ID,
+  key_secret: process.env.RAZORPAY_KEY_SECRET,
+});
+
+const transporter = nodemailer.createTransport({
+  service: 'gmail',
+  auth: {
+    user: process.env.EMAIL_USER,
+    pass: process.env.EMAIL_PASS,
+  },
+});
+
+exports.createOrder = async (req, res) => {
+  try {
+    const { amount, beneficiaryId, donorName, donorEmail, donorPhone } = req.body;
+    if (!amount || !beneficiaryId) {
+      return res.status(400).json({ success: false, error: 'Amount and beneficiary ID are required' });
+    }
+    if (typeof amount !== 'number' || amount <= 0) {
+      return res.status(400).json({ success: false, error: 'Invalid amount' });
+    }
+    const beneficiary = await Beneficiary.findById(beneficiaryId);
+    if (!beneficiary) {
+      return res.status(404).json({ success: false, error: 'Beneficiary not found' });
+    }
+    const shortBeneficiaryId = beneficiaryId.slice(0, 8);
+    const receipt = `don_${Date.now()}_${shortBeneficiaryId}`;
+    const order = await razorpay.orders.create({
+      amount: amount * 100,
+      currency: 'INR',
+      receipt,
+      notes: {
+        beneficiaryId,
+        donorName: donorName || 'Anonymous',
+        donorEmail: donorEmail || '',
+        donorPhone: donorPhone || '',
+      },
+    });
+    if (!order.id) throw new Error('Order ID not generated by Razorpay');
+    res.json({
+      success: true,
+      data: {
+        orderId: order.id,
+        amount: order.amount,
+        currency: order.currency,
+        keyId: process.env.RAZORPAY_KEY_ID,
+      },
+    });
+  } catch (error) {
+    res.status(500).json({ success: false, error: 'Failed to create payment order', details: error.message });
+  }
+};
+
+exports.verifyPayment = async (req, res) => {
+  try {
+    const { orderId, paymentId, signature, beneficiaryId, donorName, donorEmail, donorPhone, amount } = req.body;
+    if (!orderId || !paymentId || !signature) {
+      return res.status(400).json({ success: false, error: 'Missing required payment details' });
+    }
+    const generatedSignature = crypto
+      .createHmac('sha256', process.env.RAZORPAY_KEY_SECRET)
+      .update(`${orderId}|${paymentId}`)
+      .digest('hex');
+    if (generatedSignature !== signature) {
+      return res.status(400).json({ success: false, error: 'Invalid payment signature' });
+    }
+    const existingDonation = await Donation.findOne({ paymentId });
+    if (existingDonation) {
+      return res.json({ success: true, message: 'Payment already recorded' });
+    }
+    const donation = new Donation({
+      beneficiaryId,
+      amount: amount / 100,
+      donorName,
+      donorEmail,
+      donorPhone,
+      paymentId,
+      orderId,
+      paymentStatus: 'completed',
+    });
+    await donation.save();
+    const mailOptions = {
+      from: process.env.EMAIL_USER,
+      to: donorEmail,
+      subject: 'Thank You for Your Donation to Project Sitaare',
+      html: `
+        <h2>Thank You, ${donorName}!</h2>
+        <p>Your donation of ₹${(amount / 100).toLocaleString()} has been successfully received.</p>
+        <p><strong>Details:</strong></p>
+        <ul>
+          <li>Order ID: ${orderId}</li>
+          <li>Payment ID: ${paymentId}</li>
+          <li>Beneficiary: Child Education Fund</li>
+          <li>Date: ${new Date().toLocaleString()}</li>
+        </ul>
+        <p>Your support helps us empower underprivileged children through education.</p>
+        <p>Best regards,<br>Team Sitaare</p>
+      `,
+    };
+    try {
+      await transporter.sendMail(mailOptions);
+    } catch (emailError) {}
+    res.json({ success: true, message: 'Payment verified and donation recorded' });
+  } catch (error) {
+    res.status(500).json({ success: false, error: 'Failed to verify payment', details: error.message });
+  }
+};
+
+exports.paymentAuth = async (req, res) => {
+  try {
+    const { paymentId } = req.body;
+    if (!paymentId) {
+      return res.status(400).json({ success: false, error: 'Payment ID required' });
+    }
+    const payment = await razorpay.payments.fetch(paymentId);
+    if (payment.status === 'captured') {
+      res.json({ success: true, status: 'captured', payment });
+    } else if (payment.status === 'failed') {
+      res.json({ success: false, status: 'failed', error: payment.error_description });
+    } else {
+      res.json({ success: false, status: payment.status });
+    }
+  } catch (error) {
+    res.status(500).json({ success: false, error: 'Failed to fetch payment status', details: error.message });
+  }
+};
+
+exports.webhook = async (req, res) => {
+  try {
+    const webhookSecret = process.env.RAZORPAY_WEBHOOK_SECRET;
+    const signature = req.headers['x-razorpay-signature'];
+    const generatedSignature = crypto
+      .createHmac('sha256', webhookSecret)
+      .update(JSON.stringify(req.body))
+      .digest('hex');
+    if (generatedSignature !== signature) {
+      return res.status(400).json({ status: 'error', message: 'Invalid signature' });
+    }
+    const event = req.body.event;
+    if (event === 'payment.captured') {
+      const { payment, order } = req.body.payload;
+      const paymentId = payment.entity.id;
+      const orderId = order.entity.id;
+      const amount = order.entity.amount / 100;
+      const { beneficiaryId, donorName, donorEmail, donorPhone } = order.entity.notes;
+      const existingDonation = await Donation.findOne({ paymentId });
+      if (existingDonation) {
+        return res.json({ status: 'success', message: 'Duplicate payment ignored' });
+      }
+      const donation = new Donation({
+        beneficiaryId,
+        amount,
+        donorName,
+        donorEmail,
+        donorPhone,
+        paymentId,
+        orderId,
+        paymentStatus: 'completed',
+      });
+      await donation.save();
+      const mailOptions = {
+        from: process.env.EMAIL_USER,
+        to: donorEmail,
+        subject: 'Thank You for Your Donation to Project Sitaare',
+        html: `
+          <h2>Thank You, ${donorName}!</h2>
+          <p>Your donation of ₹${amount.toLocaleString()} has been successfully received.</p>
+          <p><strong>Details:</strong></p>
+          <ul>
+            <li>Order ID: ${orderId}</li>
+            <li>Payment ID: ${paymentId}</li>
+            <li>Beneficiary: Child Education Fund</li>
+            <li>Date: ${new Date().toLocaleString()}</li>
+          </ul>
+          <p>Your support helps us empower underprivileged children through education.</p>
+          <p>Best regards,<br>Team Sitaare</p>
+        `,
+      };
+      try {
+        await transporter.sendMail(mailOptions);
+      } catch (emailError) {}
+    } else if (event === 'payment.failed') {
+      const { payment, order } = req.body.payload;
+      const paymentId = payment.entity.id;
+      const orderId = order.entity.id;
+      const errorDescription = payment.entity.error_description || 'Payment failed';
+      const donation = new Donation({
+        beneficiaryId: order.entity.notes.beneficiaryId,
+        amount: order.entity.amount / 100,
+        donorName: order.entity.notes.donorName,
+        donorEmail: order.entity.notes.donorEmail,
+        donorPhone: order.entity.notes.donorPhone,
+        paymentId,
+        orderId,
+        paymentStatus: 'failed',
+        error: errorDescription,
+      });
+      await donation.save();
+    }
+    res.json({ status: 'success' });
+  } catch (error) {
+    res.status(500).json({ status: 'error', message: 'Webhook processing failed' });
+  }
+};
